+++ CHANGELOG.md
@@ -461,3 +461,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix crash from multi-word keywords on some relics (kiooeht)
 
 #### dev ####
+* Easier SpineAnimation option for custom characters (kiooeht)